# openapi.yaml
openapi: 3.1.0
info:
  title: Mattis Stats API
  version: 1.0.0
  description: |
    API for tracking player statistics for the card game "Mattis".

    **Key Features**:
    - Manage players.
    - Record game rounds and identify a single loser.
    - Grant and revoke special "Fettmattis".
    - View leaderboards for regular play (loss percentage) and Fettmattis.

    **Authentication**:
    - Operations require either a bearer token or a signed session cookie.
    - Authentication endpoints establish or revoke sessions for clients that are not yet authenticated.
    - The contract defines both a bearer token scheme for API clients and a cookie-based scheme
      for browser sessions. Use the `bearerAuth` security requirement when automating requests
      or calling protected resources from non-browser clients, and the `sessionCookie` requirement
      when an endpoint expects the signed session cookie provided by BetterAuth during interactive
      login flows.

    **Error Handling**:
    - The API uses RFC 9457 for problem details in error responses.
  contact:
    name: Kenneth Aasan
    email: kenneth@aasan.dev
servers:
  - url: /api
    x-internal: false
...
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /fettmattis:
    get:
      summary: List Fettmattis Entries
      description: Returns recent Fettmattis awards.
      operationId: listFettmattis
      tags:
        - Fettmattis
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
          description: Limit the number of Fettmattis entries returned. Defaults to 20 when omitted.
      responses:
        "200":
          description: A list of recent Fettmattis entries.
          headers:
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimit-Limit"
            RateLimit-Remaining:
              $ref: "#/components/headers/RateLimit-Remaining"
            RateLimit-Reset:
              $ref: "#/components/headers/RateLimit-Reset"
            Access-Control-Allow-Origin:
              $ref: "#/components/headers/Access-Control-Allow-Origin"
          content:
            application/json:
              schema:
                type: array
                maxItems: 200
                items:
                  $ref: "#/components/schemas/Fettmattis"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create Fettmattis
      description: Grants a Fettmattis award to a player.
      operationId: createFettmattis
      tags:
        - Fettmattis
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FettmattisCreate"
      responses:
        "201":
          description: Fettmattis created successfully.
          headers:
            RateLimit-Limit:
              $ref: "#/components/headers/RateLimit-Limit"
            RateLimit-Remaining:
              $ref: "#/components/headers/RateLimit-Remaining"
            RateLimit-Reset:
              $ref: "#/components/headers/RateLimit-Reset"
            Access-Control-Allow-Origin:
              $ref: "#/components/headers/Access-Control-Allow-Origin"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fettmattis"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    Player:
      type: object
      required:
        - id
        - display_name
        - active
      properties:
        id:
          type: string
          format: uuid
          maxLength: 36
          description: UUID v7, time-ordered identifier.
        display_name:
          type: string
          maxLength: 64
          pattern: "^.{1,64}$"
        active:
          type: boolean
...
    Fettmattis:
      type: object
      required:
        - id
        - player
        - created_at
      properties:
        id:
          type: string
          format: uuid
          maxLength: 36
          description: UUID v7, time-ordered identifier.
        player:
          $ref: "#/components/schemas/Player"
        round_id:
          type: string
          format: uuid
          maxLength: 36
          description: UUID v7, time-ordered identifier.
          nullable: true
        created_at:
          type: string
          format: date-time
          maxLength: 30
    FettmattisCreate:
      type: object
      required:
        - player_id
      properties:
        player_id:
          type: string
          format: uuid
          maxLength: 36
          description: UUID v7, time-ordered identifier.
        round_id:
          type: string
          format: uuid
          maxLength: 36
          description: UUID v7, time-ordered identifier.
          nullable: true
...
    Problem:
      type: object
      description: RFC 9457 Problem Details
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          maxLength: 2048
          default: "about:blank"
        title:
          type: string
          maxLength: 200
          pattern: "^.{1,200}$"
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
        detail:
          type: string
          maxLength: 2000
          pattern: "^.{1,2000}$"
        instance:
          type: string
          format: uri
          maxLength: 2048
  responses:
    BadRequest:
      description: The request body is invalid.
      headers:
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimit-Limit"
        RateLimit-Remaining:
          $ref: "#/components/headers/RateLimit-Remaining"
        RateLimit-Reset:
          $ref: "#/components/headers/RateLimit-Reset"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Authentication is required for this operation.
      headers:
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimit-Limit"
        RateLimit-Remaining:
          $ref: "#/components/headers/RateLimit-Remaining"
        RateLimit-Reset:
          $ref: "#/components/headers/RateLimit-Reset"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Conflict:
      description: The resource already exists or the request conflicts with the current state.
      headers:
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimit-Limit"
        RateLimit-Remaining:
          $ref: "#/components/headers/RateLimit-Remaining"
        RateLimit-Reset:
          $ref: "#/components/headers/RateLimit-Reset"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: The client has sent too many requests in a given amount of time.
      headers:
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimit-Limit"
        RateLimit-Remaining:
          $ref: "#/components/headers/RateLimit-Remaining"
        RateLimit-Reset:
          $ref: "#/components/headers/RateLimit-Reset"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InternalServerError:
      description: The server encountered an unexpected error.
      headers:
        RateLimit-Limit:
          $ref: "#/components/headers/RateLimit-Limit"
        RateLimit-Remaining:
          $ref: "#/components/headers/RateLimit-Remaining"
        RateLimit-Reset:
          $ref: "#/components/headers/RateLimit-Reset"
        Access-Control-Allow-Origin:
          $ref: "#/components/headers/Access-Control-Allow-Origin"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  headers:
    Access-Control-Allow-Origin:
      description: Allowed origin for cross-origin requests.
      schema:
        type: string
        maxLength: 2048
        pattern: "^\\S+$"
    RateLimit-Limit:
      description: The maximum number of requests permitted in the current window.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 10000
    RateLimit-Remaining:
      description: The number of remaining requests in the current window.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 10000
    RateLimit-Reset:
      description: Seconds remaining until the rate limit window resets.
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 86400
    Retry-After:
      description: Seconds to wait before making a new request.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 86400
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Session
      description: >-
        Session token returned from the authentication endpoints. Present the
        value in the `Authorization` header using the `Bearer` scheme for
        protected write operations or when automating calls from services that
        cannot store cookies.
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: >-
        Signed session cookie issued during authentication. Present this cookie
        when invoking interactive session endpoints (e.g., inspect or sign out)
        from user agents that manage browser storage. In secure environments the
        cookie is prefixed with `__Secure-`.